refactor: Renamed AnnotationTable to AnnotationDisplayTable
